+++ R/is_dtx.R
@@ -1,7 +1,7 @@
 #' Test if the object is a data frame (data.frame, data.table or tibble)
 #'
 #' @param x An object
-#' @param strict Should this be strictly the corresponding class `TRUE`, by default, or could it be subclassed too (`FALSE`).
+#' @param strict Should this be strictly the corresponding class `TRUE`, by default, or could it be subclassed too (`FALSE`). With `strict = TRUE`, the **grouped_df** tibbles and **grouped_ts** tsibbles are also considered (tibbles or tsibbles where [dplyr::group_by()] was applied).
 #'
 #' @return These functions return `TRUE` if the object is of the correct class, otherwise they return `FALSE`. `is_dtx()` return `TRUE` if `x` is one of a data.frame, data.table or tibble.
 #' @export
@@ -17,13 +17,18 @@
 #' is_dtx(as_dtt(mtcars)) # TRUE
 #' is_dtbl(as_dtbl(mtcars)) # TRUE
 #' is_dtx(as_dtbl(mtcars)) # TRUE
+#' is_dtx(as_dtbl(mtcars) |> dplyr::group_by(cyl)) # TRUE (special case)
 #'
 #' is_dtx("some string") # FALSE
 is_dtx <- function(x, strict = TRUE) {
   if (isTRUE(strict)) {
-    # One of the three type, and not subclassed
+    # One of the three type, and not subclassed (except with grouped_df)
     class1 <- class(x)[1]
-    class1 == "data.frame" || class1 == "data.table" || class1 == "tbl_df"
+    if (class1 == "grouped_df" || class1 == "grouped_ts") {# Special case
+      inherits(x, "tbl_df")
+    } else {
+      class1 == "data.frame" || class1 == "data.table" || class1 == "tbl_df"
+    }
   } else {
     # All three types inherits from data.frame
     inherits(x, "data.frame")
@@ -54,7 +59,12 @@ is_dtt <- function(x, strict = TRUE) {
 #' @rdname is_dtx
 is_dtbl <- function(x, strict = TRUE) {
   if (isTRUE(strict)) {
-    class(x)[1] == "tbl_df"
+    class1 <- class(x)[1]
+    if (class1 == "grouped_df" || class1 == "grouped_ts") {# Special case
+      inherits(x, "tbl_df")
+    } else {
+      class1 == "tbl_df"
+    }
   } else {
     inherits(x, "tbl_df")
   }