+++ R/svBase-package.R
@@ -14,7 +14,7 @@
 
 #' @importFrom data.table as.data.table haskey key setDT setkeyv
 #' @importFrom tibble as_tibble tibble
-#' @importFrom dplyr collect
+#' @importFrom dplyr collect ungroup
 #' @importFrom pillar tbl_sum
 #' @import zeallot
 # The following block is used by usethis to automatically manage